# server/openapi.yaml
openapi: 3.0.3
info:
  title: Grasp Platform API (partial)
  version: 0.1.0
  description: |
    Minimal OpenAPI contract for platform list endpoints.

    **Pagination pattern (shared):** GET /api/artifacts, GET /api/runs,
    GET /api/gates and GET /api/projects/{id}/audit always return the paginated
    envelope `{ items, total, page, pageSize, hasMore }` as implemented in
    `internal/handlers/pagination.go` (`parsePagination`, `paginatedResponse`).
    Omitted `page` / `pageSize` default to `1` / `20`; `maxPageSize=100`;
    `hasMore = page × pageSize < total`.
servers:
  - url: /api
    description: Platform HTTP API (session cookie or Bearer workflow API key)
...
paths:
...
  /artifacts:
    get:
      operationId: listArtifacts
      tags: [artifacts]
      summary: List platform artifacts (paginated)
      description: |
        Returns artifacts across all runs for the Artifacts view and integrations,
        always as the paginated envelope `{ items, total, page, pageSize, hasMore }`.

        **Pagination unit:**
        - Default (no `groupBy`, or any value other than `run`): `pageSize` /
          `total` count **artifact rows**; `items` is a slice of artifacts.
        - `groupBy=run` (opt-in, ArtifactsView L2): `pageSize` / `total` count
          **Runs**; `items` is the flat list of **all** artifacts belonging to
          the Runs on this page (whole-Run expansion). Therefore `items.length`
          may be greater than `pageSize` when a Run has many artifacts.
          Runs are ordered by `MAX(artifacts.created_at)` descending.
          When `q` is set, a Run qualifies if any of its artifacts match; the
          expand step drops `q` so non-matching files in that Run are still
          returned (wf / project filters still apply).

        **Filtering:**
        - `wf`: restrict to a workflow ID, or `__unnamed__` for artifacts whose
          `workflow_id` is NULL or empty string.
        - `q`: case-insensitive substring search across `artifacts.name`,
          `artifacts.node_id`, and `runs.title` (via LEFT JOIN).

        **List item shape:** `items` omit heavy `content`; `runTitle` is
        populated from `runs.title` when non-empty. Empty results normalize
        `items` to `[]` (never JSON `null`).

        Pagination validation matches `/api/runs` and `/api/gates`: invalid
        `page`/`pageSize` or `pageSize > 100` → HTTP 400.
      parameters:
        - $ref: '#/components/parameters/PageQuery'
        - $ref: '#/components/parameters/PageSizeQuery'
        - $ref: '#/components/parameters/WfQuery'
        - $ref: '#/components/parameters/ArtifactSearchQuery'
        - $ref: '#/components/parameters/ArtifactGroupByQuery'
      responses:
        '200':
          description: Paginated artifact list
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ArtifactPaginatedResponse'
              examples:
                paginated:
                  summary: Default — by artifact
                  value:
                    items:
                      - id: art-abc123
                        runId: run-1
                        nodeId: plan
                        workflowId: wf-product
                        name: plan.json
                        kind: json
                        sizeBytes: 1024
                        createdAt: '2026-07-12T00:00:00Z'
                        runTitle: Sprint Planning
                    total: 47
                    page: 1
                    pageSize: 20
                    hasMore: true
                paginatedByRun:
                  summary: groupBy=run (L2 — by Run; items may exceed pageSize)
                  value:
                    items:
                      - id: art-abc123
                        runId: run-1
                        nodeId: plan
                        workflowId: wf-product
                        name: plan.json
                        kind: json
                        sizeBytes: 1024
                        createdAt: '2026-07-12T00:00:00Z'
                        runTitle: Sprint Planning
                      - id: art-def456
                        runId: run-1
                        nodeId: research
                        workflowId: wf-product
                        name: research.json
                        kind: json
                        sizeBytes: 2048
                        createdAt: '2026-07-12T00:01:00Z'
                        runTitle: Sprint Planning
                    total: 12
                    page: 1
                    pageSize: 20
                    hasMore: false
                emptyPage:
                  summary: No matches
                  value:
                    items: []
                    total: 0
                    page: 1
                    pageSize: 20
                    hasMore: false
        '400':
          description: Invalid pagination parameters
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
              examples:
                pageSizeTooLarge:
                  value:
                    error: pageSize must be ≤ 100
        '401':
          description: Missing or invalid session / API key
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
...
components:
  parameters:
    PageQuery:
      name: page
      in: query
      required: false
      description: |
        1-based page index. Defaults to `1`.
        Must be a positive integer (≥ 1).
      schema:
        type: integer
        minimum: 1
        default: 1
    PageSizeQuery:
      name: pageSize
      in: query
      required: false
      description: |
        Page size. Defaults to `20`.
        Maximum `100`; values above 100 return HTTP 400.
      schema:
        type: integer
        minimum: 1
        maximum: 100
        default: 20
    WfQuery:
      name: wf
      in: query
      required: false
      description: |
        Filter by workflow ID. Use the sentinel value `__unnamed__` to list
        artifacts with no workflow (`workflow_id` IS NULL or empty).
      schema:
        type: string
        example: wf-product
    ArtifactSearchQuery:
      name: q
      in: query
      required: false
      description: |
        Case-insensitive substring search against artifact `name`, `nodeId`,
        and associated run `title` (SQL `LOWER LIKE` with `%` wildcards).
        With `groupBy=run`, `q` selects which Runs qualify; expansion returns
        the full Run (see operation description).
      schema:
        type: string
        example: plan
    ArtifactGroupByQuery:
      name: groupBy
      in: query
      required: false
      description: |
        Pagination grouping mode.
        Omit or use any value other than `run` to page by artifact row
        (`total` / `pageSize` = artifact count).
        Set to `run` to page by Run (`total` / `pageSize` = Run count;
        `items` = flat whole-Run artifact lists for the page).
      schema:
        type: string
        enum: [run]
        example: run
  schemas:
    Error:
      type: object
      properties:
        error:
          type: string
      required: [error]
...
    ArtifactListItem:
      type: object
      description: |
        Lightweight artifact metadata for list endpoints. Does **not** include
        `content` (use GET /artifacts/{id}/content for full body).
      properties:
        id:
          type: string
        runId:
          type: string
        nodeId:
          type: string
        workflowId:
          type: string
        workflowName:
          type: string
        name:
          type: string
        kind:
          type: string
        sizeBytes:
          type: integer
        createdAt:
          type: string
          format: date-time
        runTitle:
          type: string
          description: From LEFT JOIN runs.title; omitted when empty
      required: [id, runId, nodeId, name, kind, sizeBytes, createdAt]
    PaginatedResponseBase:
      type: object
      description: |
        Shared pagination envelope used by GET /api/artifacts, GET /api/runs,
        GET /api/gates and GET /api/projects/{id}/audit (see pagination.go).
      properties:
        total:
          type: integer
          description: Total matching rows before paging
        page:
          type: integer
          description: Current 1-based page
        pageSize:
          type: integer
          description: Requested page size (default 20, max 100)
        hasMore:
          type: boolean
          description: True when `page × pageSize < total`
      required: [total, page, pageSize, hasMore]
    ArtifactPaginatedResponse:
      allOf:
        - $ref: '#/components/schemas/PaginatedResponseBase'
        - type: object
          properties:
            items:
              type: array
              items:
                $ref: '#/components/schemas/ArtifactListItem'
              description: Always an array; empty results are `[]` not `null`
          required: [items]